==> 1. Preprocess_promcse_train_patent_pos_neg_cross_ctx.ipynb <==
# Prepare Patent dataset to be used for PromCSE Training

import re
import os
import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

!ls /content/drive/

PROJECT_DIR = "/content/drive/MyDrive/patent"

DATA_DIR = PROJECT_DIR + '/data'

BERT_SEP = " [SEP] "
#BERT_SEP = " . "
TARGET_SEP = " ; "
DF_SEP = "|"

# Join CPC text by their code

def get_cpc_texts():
  contexts = []
  pattern = '[A-Z]\d+'
  for file_name in os.listdir(os.path.join(DATA_DIR, 'cpc-data/CPCSchemeXML202205')):
    result = re.findall(pattern, file_name)
    if result:
      contexts.append(result)
  contexts = sorted(set(sum(contexts, [])))
  results = {}

  for cpc in ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'Y']:
    with open(os.path.join(DATA_DIR, f'cpc-data/CPCTitleList202205/cpc-section-{cpc}_20220501.txt'
    )) as f:
      s = f.read()
    pattern = f'{cpc}\t\t.+'
    result = re.findall(pattern, s)
    pattern = "^" + pattern[:-2]
    cpc_result = re.sub(pattern, "", result[0])
    for context in [c for c in contexts if c[0] == cpc]:
      pattern = f'{context}\t\t.+'
      result = re.findall(pattern, s)
      pattern = "^" + pattern[:-2]
      sep = BERT_SEP
      #results[context] = cpc_result + sep + re.sub(pattern, "", result[0])
      results[context] = cpc_result.lower() + sep + re.sub(pattern, "", result[0]).lower()
      #results[context] = results[context].lower()
  return results

cpc_texts = get_cpc_texts()

# Patent Data

df = pd.read_csv(DATA_DIR + "/patent_orig/train.csv")

df.shape

df.columns

df.info()

df.head(10)

# Positive targets are with score = 1, 0.75, 0.5 and Negative targets are with score = 0, 0.25

df_pos = df[df.score >= 0.5 ]
df_pos.shape

df_neg = df[df.score < 0.5 ]
df_neg.shape

# Join positive targets. Create new row with this positive target

df_pos.head(20)

agg_pos = df_pos.groupby(['anchor', 'context'])['target'].apply(list).reset_index(name = 'target')


agg_pos['target'] =  agg_pos['target'].apply(lambda l: TARGET_SEP.join(l))

agg_pos['score']=0.50

agg_pos.head(10)

df_pos = pd.concat([df_pos, agg_pos])

df_pos.shape

df_pos.loc[(df_pos['context']=='A61') & (df_pos['anchor']=='abatement')].head(20)

df_pos['sen t1'] = df_pos['target']

df_pos.drop(['id', 'target', 'score'], axis=1, inplace=True)

df_pos.drop_duplicates(inplace=True)

# Join negative targets. Add to hard negative for that context and anchor.

df_neg.head()

agg_neg = df_neg.groupby(['anchor', 'context'])['target'].apply(list).reset_index(name = 'target')
agg_neg['target'] =  agg_neg['target'].apply(lambda l: TARGET_SEP.join(l))
agg_neg['score']=0.00

agg_neg.head(10)

df_neg = pd.concat([df_neg, agg_neg])

df_neg.shape

df_neg.loc[(df_neg['context']=='A61') & (df_neg['anchor']=='abatement')].head(20)

df_neg['hard_neg'] = df_neg['target']

df_neg.drop(['id', 'target', 'score'], axis=1, inplace=True)

df_neg.drop_duplicates(inplace=True)

df_pos.head()

df_pos.shape

df_neg.head()

df_neg.shape

# Negative samples are negative for that context, may be positive for other contexts

df_pos['key'] = 1
df_neg['key'] = 1

df_all = pd.merge(df_pos, df_neg, on =['key', 'anchor', 'context' ]).drop("key", 1)

df_all.shape

df_all.head(20)

# Augment with CPC data

df_all['context_text'] = df_all['context'].map(cpc_texts)

# SimCSE Format

df_all['sent0'] = df_all['anchor'] + BERT_SEP + df_all['context_text']

df_all.head()

df_all.drop(['anchor', 'context', 'context_text' ], axis=1, inplace=True)

df_all.shape

df_all.to_csv(DATA_DIR + '/patent_train_simcse.csv', sep=DF_SEP, index=False)
