==> 2. Preprocess_promcse_train_patent_snli.ipynb <==
# Merge Patent and SNLI dataset for Training using PromCSE

import re
import os
import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

!ls /content/drive/

PROJECT_DIR = "/content/drive/MyDrive/patent"

DATA_DIR = PROJECT_DIR + '/data'

BERT_SEP = " [SEP] "
#BERT_SEP = " . "
TARGET_SEP = " ; "
DF_SEP = "|"

# Merge SNLI and Patent dataset

patent_df = pd.read_csv(DATA_DIR + '/patent_train_simcse.csv', sep = DF_SEP)

patent_df.shape

patent_df = patent_df[['sent0', 'sent1', 'hard_neg']]

patent_df.head()

snli_df = pd.read_csv(DATA_DIR + '/nli_for_simcse.csv')

snli_df.shape

snli_df.head()

all_df = pd.concat([patent_df, snli_df], axis=0, ignore_index = True)

all_df.shape

all_df.head()

from sklearn.utils import shuffle
shuff_df = shuffle(all_df)

shuff_df.shape

shuff_df.head()

shuff_df.to_csv(DATA_DIR + '/patent_snli.csv', index=False, sep = DF_SEP)
